Add docker client to images
The `gcloud docker` subcommand relies on executing a docker client.
This change allows the user to bindmount in the host's docker socket,
and then interact with it using the `gcloud docker` subcommand.

This will only work if the user's host dockerd is compatible with this
docker client, but this is true for all modern versions of docker.

See additional discussion in https://github.com/GoogleCloudPlatform/cloud-sdk-docker/issues/97
for why bindmounting in the client's docker binary is non-ideal.

This change is the equivilant of simply downloading their official linux
release tarball, and picking the client binary out of it.

This is simply a more readable, and possibly more future-proof way, of
accomplishing that.