Add a 'gcloud --version' failsafe to docker builds
This adds a "the binary runs" check to the end of each container's list
of commands.

The primary benefit this has is that a build will fail if any of those
commands don't work, which is preferable to a bad release being made.

As a small secondary benefit, it also includes the `version` output in
the build log.